Exit manager add requests.rs and utils.rs
This is part of a continued effort to reduce the huge size of these
files and get something more readable out of this module.
