+++ rita_client/src/exit_manager/mod.rs
@@ -19,33 +19,16 @@
 pub mod encryption;
 pub mod exit_loop;
 pub mod exit_switcher;
+pub mod requests;
 pub mod time_sync;
+pub mod utils;
 
-use crate::heartbeat::get_exit_registration_state;
-use crate::heartbeat::get_selected_exit_server;
-use crate::rita_loop::CLIENT_LOOP_TIMEOUT;
-use crate::RitaClientError;
-use actix_web_async::Result;
-use althea_kernel_interface::{
-    exit_client_tunnel::ClientExitTunnelConfig, DefaultRoute, KernelInterfaceError,
-};
-use althea_types::exit_identity_to_id;
-use althea_types::ExitClientDetails;
-use althea_types::ExitDetails;
 use althea_types::ExitListV2;
-use althea_types::WgKey;
-use althea_types::{ExitClientIdentity, ExitRegistrationDetails, ExitState};
-use babel_monitor::structs::Route;
-use encryption::decrypt_exit_list;
-use encryption::decrypt_exit_state;
-use encryption::encrypt_exit_client_id;
+use althea_types::ExitState;
 use exit_switcher::ExitTracker;
-use ipnetwork::IpNetwork;
-use rita_common::KI;
-use settings::client::{ExitServer, SelectedExit};
-use settings::set_rita_client;
+use settings::client::SelectedExit;
 use std::collections::HashMap;
-use std::net::{IpAddr, SocketAddr};
+use std::net::IpAddr;
 use std::sync::Arc;
 use std::sync::RwLock;
 use std::time::Instant;
@@ -133,398 +116,12 @@ pub fn set_selected_exit(exit_info: SelectedExit) {
     SELECTED_EXIT_DETAILS.write().unwrap().selected_exit = exit_info;
 }
 
-fn linux_setup_exit_tunnel(
-    general_details: &ExitDetails,
-    our_details: &ExitClientDetails,
-) -> Result<(), RitaClientError> {
-    let mut rita_client = settings::get_rita_client();
-    let mut network = rita_client.network;
-    let local_mesh_ip = network.mesh_ip;
-
-    // TODO this should be refactored to return a value
-    KI.update_settings_route(&mut network.last_default_route)?;
-    info!("Updated settings route");
-
-    if let Err(KernelInterfaceError::RuntimeError(v)) = KI.create_blank_wg_interface("wg_exit") {
-        return Err(RitaClientError::MiscStringError(v));
-    }
-
-    let selected_exit = get_selected_exit_server().expect("There should be a selected exit here");
-    let args = ClientExitTunnelConfig {
-        endpoint: SocketAddr::new(
-            selected_exit.exit_id.mesh_ip,
-            selected_exit.wg_exit_listen_port,
-        ),
-        pubkey: selected_exit.exit_id.wg_public_key,
-        private_key_path: network.wg_private_key_path.clone(),
-        listen_port: DEFAULT_WG_LISTEN_PORT,
-        local_ip: our_details.client_internal_ip,
-        netmask: general_details.netmask,
-        rita_hello_port: network.rita_hello_port,
-        user_specified_speed: network.user_bandwidth_limit,
-    };
-
-    info!("Args while setting up wg_exit on client are: {:?}", args);
-
-    rita_client.network = network;
-    settings::set_rita_client(rita_client);
-
-    KI.set_client_exit_tunnel_config(args, local_mesh_ip)?;
-    KI.set_route_to_tunnel(&general_details.server_internal_ip)?;
-    KI.set_ipv6_route_to_tunnel()?;
-
-    KI.create_client_nat_rules()?;
-
-    Ok(())
-}
-
-fn restore_nat() {
-    if let Err(e) = KI.restore_client_nat() {
-        error!("Failed to restore client nat! {:?}", e);
-    }
-}
-
-fn remove_nat() {
-    if let Err(e) = KI.block_client_nat() {
-        error!("Failed to block client nat! {:?}", e);
-    }
-}
-
-/// When we retrieve an exit list from an exit, add the compatible exits to the exit server list.
-/// This allows these exits to move to GotInfo state, allowing us to switch or connect quickly
-pub fn add_exits_to_exit_server_list(list: ExitListV2) {
-    let mut rita_client = settings::get_rita_client();
-    let mut exits = rita_client.exit_client.bootstrapping_exits;
-
-    for e in list.exit_list {
-        exits.entry(e.mesh_ip).or_insert(ExitServer {
-            exit_id: exit_identity_to_id(e.clone()),
-            registration_port: e.registration_port,
-            wg_exit_listen_port: e.wg_exit_listen_port,
-        });
-    }
-
-    // Update settings with new exits
-    rita_client.exit_client.bootstrapping_exits = exits;
-    set_rita_client(rita_client);
-}
-
-async fn send_exit_setup_request(
-    exit_pubkey: WgKey,
-    to: SocketAddr,
-    ident: ExitClientIdentity,
-) -> Result<ExitState, RitaClientError> {
-    let endpoint = format!("http://[{}]:{}/secure_setup", to.ip(), to.port());
-
-    let ident = encrypt_exit_client_id(&exit_pubkey.into(), ident);
-
-    let client = awc::Client::default();
-
-    let response = client
-        .post(&endpoint)
-        .timeout(CLIENT_LOOP_TIMEOUT)
-        .send_json(&ident)
-        .await;
-    let mut response = match response {
-        Ok(a) => a,
-        Err(awc::error::SendRequestError::Timeout) => {
-            return Err(RitaClientError::SendRequestError(
-                awc::error::SendRequestError::Timeout.to_string(),
-            ));
-        }
-        Err(e) => return Err(RitaClientError::SendRequestError(e.to_string())),
-    };
-
-    let value = response.json().await?;
-
-    match decrypt_exit_state(value, exit_pubkey.into()) {
-        Err(e) => Err(e),
-        a => a,
-    }
-}
-
-async fn send_exit_status_request(
-    exit_pubkey: WgKey,
-    to: &SocketAddr,
-    ident: ExitClientIdentity,
-) -> Result<ExitState, RitaClientError> {
-    let endpoint = format!("http://[{}]:{}/secure_status", to.ip(), to.port());
-    let ident = encrypt_exit_client_id(&exit_pubkey.into(), ident);
-
-    let client = awc::Client::default();
-    let response = client
-        .post(&endpoint)
-        .timeout(CLIENT_LOOP_TIMEOUT)
-        .send_json(&ident)
-        .await;
-
-    let mut response = match response {
-        Ok(a) => a,
-        Err(awc::error::SendRequestError::Timeout) => {
-            // Did not get a response, is it a rogue exit or some netork error?
-            return Err(RitaClientError::SendRequestError(
-                awc::error::SendRequestError::Timeout.to_string(),
-            ));
-        }
-        Err(e) => return Err(RitaClientError::SendRequestError(e.to_string())),
-    };
-    let value = response.json().await?;
-
-    match decrypt_exit_state(value, exit_pubkey.into()) {
-        Err(e) => Err(e),
-        Ok(a) => Ok(a),
-    }
-}
-
-/// Registration is simply one of the exits requesting an update to a global smart contract
-/// with our information.
-pub async fn exit_setup_request(code: Option<String>) -> Result<(), RitaClientError> {
-    let client_settings = settings::get_rita_client();
-    let exit =
-        match get_selected_exit_server() {
-            Some(exit) => exit,
-            None => return Err(RitaClientError::MiscStringError(
-                "Not yet connected to any exit to setup, please check connectivity and try again"
-                    .to_string(),
-            )),
-        };
-
-    match client_settings.exit_client.registration_state {
-        ExitState::New { .. } | ExitState::Pending { .. } => {
-            let exit_pubkey = exit.exit_id.wg_public_key;
-
-            let mut reg_details: ExitRegistrationDetails =
-                match client_settings.payment.contact_info {
-                    Some(val) => val.into(),
-                    None => {
-                        return Err(RitaClientError::MiscStringError(
-                            "No registration info set!".to_string(),
-                        ))
-                    }
-                };
-
-            // Send a verification code if we have one
-            reg_details.phone_code = code;
-
-            let ident = ExitClientIdentity {
-                global: match settings::get_rita_client().get_identity() {
-                    Some(id) => id,
-                    None => {
-                        return Err(RitaClientError::MiscStringError(
-                            "Identity has no mesh IP ready yet".to_string(),
-                        ));
-                    }
-                },
-                wg_port: DEFAULT_WG_LISTEN_PORT,
-                reg_details,
-            };
-
-            let endpoint = SocketAddr::new(exit.exit_id.mesh_ip, exit.registration_port);
-
-            info!(
-                "sending exit setup request {:?} to {:?}, using {:?}",
-                ident, exit, endpoint
-            );
-
-            let exit_response = send_exit_setup_request(exit_pubkey, endpoint, ident).await?;
-
-            info!("Setting an exit setup response");
-            // we already have a loaded rita client settings above, but it could have been several seconds
-            // since we loaded above, better to load a new copy just in case
-            let mut client_settings = settings::get_rita_client();
-            client_settings.exit_client.registration_state = exit_response;
-            set_rita_client(client_settings);
-
-            return Ok(());
-        }
-        ExitState::Denied { message } => {
-            warn!(
-                "Exit {} is in ExitState DENIED with {}, not able to be setup",
-                exit.exit_id.mesh_ip, message
-            );
-        }
-        ExitState::Registered { .. } => {
-            warn!(
-                "Exit {} already reports us as registered",
-                exit.exit_id.mesh_ip
-            )
-        }
-    }
-
-    Err(RitaClientError::MiscStringError(
-        "Could not find a valid exit to register to!".to_string(),
-    ))
-}
-
-async fn exit_status_request(exit: IpAddr) -> Result<(), RitaClientError> {
-    let current_exit = match settings::get_rita_client()
-        .exit_client
-        .bootstrapping_exits
-        .get(&exit)
-    {
-        Some(current_exit) => current_exit.clone(),
-        None => {
-            return Err(RitaClientError::NoExitError(exit.to_string()));
-        }
-    };
-    let reg_details = match settings::get_rita_client().payment.contact_info {
-        Some(val) => val.into(),
-        None => {
-            return Err(RitaClientError::MiscStringError(
-                "No valid details".to_string(),
-            ))
-        }
-    };
-
-    let exit_pubkey = current_exit.exit_id.wg_public_key;
-    let ident = ExitClientIdentity {
-        global: match settings::get_rita_client().get_identity() {
-            Some(id) => id,
-            None => {
-                return Err(RitaClientError::MiscStringError(
-                    "Identity has no mesh IP ready yet".to_string(),
-                ));
-            }
-        },
-        wg_port: DEFAULT_WG_LISTEN_PORT,
-        reg_details,
-    };
-
-    let endpoint = SocketAddr::new(current_exit.exit_id.mesh_ip, current_exit.registration_port);
-
-    trace!(
-        "sending exit status request to {} using {:?}",
-        exit,
-        endpoint
-    );
-
-    let exit_response = send_exit_status_request(exit_pubkey, &endpoint, ident).await?;
-    let mut rita_client = settings::get_rita_client();
-    rita_client.exit_client.registration_state = exit_response.clone();
-    settings::set_rita_client(rita_client);
-
-    trace!("Got exit status response {:?}", exit_response);
-    Ok(())
-}
-
-/// Hits the exit_list endpoint for a given exit.
-async fn get_exit_list(exit: IpAddr) -> Result<ExitListV2, RitaClientError> {
-    let current_exit = match settings::get_rita_client()
-        .exit_client
-        .bootstrapping_exits
-        .get(&exit)
-    {
-        Some(current_exit) => current_exit.clone(),
-        None => {
-            return Err(RitaClientError::NoExitError(exit.to_string()));
-        }
-    };
-
-    let exit_pubkey = current_exit.exit_id.wg_public_key;
-    let reg_details = match settings::get_rita_client().payment.contact_info {
-        Some(val) => val.into(),
-        None => {
-            return Err(RitaClientError::MiscStringError(
-                "No valid details".to_string(),
-            ))
-        }
-    };
-    let ident = ExitClientIdentity {
-        global: match settings::get_rita_client().get_identity() {
-            Some(id) => id,
-            None => {
-                return Err(RitaClientError::MiscStringError(
-                    "Identity has no mesh IP ready yet".to_string(),
-                ));
-            }
-        },
-        wg_port: DEFAULT_WG_LISTEN_PORT,
-        reg_details,
-    };
-
-    let exit_server = current_exit.exit_id.mesh_ip;
-
-    let endpoint = format!(
-        "http://[{}]:{}/exit_list_v2",
-        exit_server, current_exit.registration_port
-    );
-    let ident = encrypt_exit_client_id(&exit_pubkey.into(), ident);
-
-    let client = awc::Client::default();
-    let response = client
-        .post(&endpoint)
-        .timeout(CLIENT_LOOP_TIMEOUT)
-        .send_json(&ident)
-        .await;
-    let mut response = match response {
-        Ok(a) => a,
-        Err(awc::error::SendRequestError::Timeout) => {
-            // Did not get a response, is it a rogue exit or some netork error?
-            return Err(RitaClientError::SendRequestError(
-                awc::error::SendRequestError::Timeout.to_string(),
-            ));
-        }
-        Err(e) => return Err(RitaClientError::SendRequestError(e.to_string())),
-    };
-
-    let value = response.json().await?;
-
-    match decrypt_exit_list(value, exit_pubkey.into()) {
-        Err(e) => Err(e),
-        Ok(a) => Ok(a),
-    }
-}
-
-fn correct_default_route(input: Option<DefaultRoute>) -> bool {
-    match input {
-        Some(v) => v.is_althea_default_route(),
-        None => false,
-    }
-}
-
-/// This function takes a list of babel routes and uses this to insert ip -> route
-/// instances in the hashmap. This is an optimization that allows us to reduce route lookups from O(n * m ) to O(m + n)
-/// when trying to find exit ips in our cluster
-fn get_routes_hashmap(routes: Vec<Route>) -> HashMap<IpAddr, Route> {
-    let mut ret = HashMap::new();
-    for r in routes {
-        ret.insert(r.prefix.ip(), r);
-    }
-
-    ret
-}
-
-pub fn get_client_pub_ipv6() -> Option<IpNetwork> {
-    let exit_info = get_exit_registration_state();
-    if let ExitState::Registered { our_details, .. } = exit_info {
-        return our_details.internet_ipv6_subnet;
-    }
-    None
-}
-
-/// Verfies if exit has changed to reestablish wg tunnels
-/// 1.) When exit instance ip has changed
-/// 2.) Exit reg details have chaged
-pub fn has_exit_changed(
-    last_states: LastExitStates,
-    selected_exit: IpAddr,
-    current_reg_state: ExitState,
-) -> bool {
-    let last_exit = last_states.last_exit;
-
-    let instance_has_changed = !(last_exit.is_some() && last_exit.unwrap() == selected_exit);
-
-    let last_exit_details = last_states.last_exit_details;
-    let exit_reg_has_changed =
-        !(last_exit_details.is_some() && last_exit_details.unwrap() == current_reg_state);
-
-    instance_has_changed | exit_reg_has_changed
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use althea_types::{ExitVerifMode, Identity, SystemChain};
+    use althea_types::{ExitClientDetails, ExitDetails, ExitVerifMode, Identity, SystemChain};
+    use settings::client::ExitServer;
+    use utils::has_exit_changed;
 
     #[test]
     fn test_exit_has_changed() {
